requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# CrewAI framework
crewai>=1.8.0
crewai-tools>=0.17.0

# Anthropic Claude API
anthropic==0.39.0

# Supabase client
supabase>=2.12.0

# Database
asyncpg>=0.30.0
sqlalchemy==2.0.36

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Input validation & sanitization
bleach==6.2.0
pydantic>=2.11.9
pydantic-settings>=2.10.1

# Rate limiting
slowapi==0.1.9

# Environment variables
python-dotenv>=1.1.1

# HTTP client
httpx==0.27.2

# Utilities
python-dateutil==2.9.0
